pyproject.toml
[tool.poetry]
name = "test_with_http_frontend"
version = "0.0.1"
description = ""
authors = ["Takaaki Mizuno <takaaki.mizuno@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.11"
podder-task-foundation = {git = "git@github.com:podder-ai/podder-task-foundation.git", rev = "0.2.14"}
"pdfminer.six" = "^20221105"
podder-task-foundation-commands-http = {path = "../../podder-task-foundation-commands-http"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
isort = "^5.10.1"
yapf = "^0.31.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
